Move cfg test to doc to avoid feature testing the whole test suite
