+++ doc/cfg/Cargo.toml
@@ -1,0 +1,19 @@
+[package]
+name = "cfg"
+version = "0.21.0"
+authors = ["Mathis Brossier @k2d222"]
+workspace = "../.." # <-- We added this and everything after!
+edition = "2021"
+
+[build-dependencies]
+lalrpop = { version = "0.21.0", path = "../../lalrpop" }
+
+[dependencies]
+lalrpop-util = { version = "0.21.0", path = "../../lalrpop-util", features = [
+    "lexer",
+] }
+
+[features]
+default = ["test-set"]
+test-set = []
+test-not-set = []
