+++ doc/cfg/build.rs
@@ -1,0 +1,3 @@
+fn main() {
+    lalrpop::process_src().unwrap();
+}
